+++ src/OpenSage.Game/Diagnostics/LuaScriptConsole.cs
@@ -1,10 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Numerics;
 using ImGuiNET;
-using OpenSage.Diagnostics.Util;
 using MoonSharp.Interpreter;
-using OpenSage.Scripting;
 
 namespace OpenSage.Diagnostics
 {
@@ -14,11 +11,9 @@ namespace OpenSage.Diagnostics
         public static string _scriptConsoleTextAll = "";
         public static Vector4 _consoleTextColor = new Vector4(0, 150, 0, 1);
 
-        public override Vector2 DefaultSize { get; } = new Vector2(150, 50);
-
         public override string DisplayName { get; } = "LUA script console";
 
-        public LuaScriptConsole(DiagnosticViewContext context) : base(context){}
+        public LuaScriptConsole(DiagnosticViewContext context) : base(context) { }
 
         protected override void DrawOverride(ref bool isGameViewFocused)
         {
@@ -26,33 +21,28 @@ namespace OpenSage.Diagnostics
             ImGui.InputTextMultiline("", ref _scriptConsoleText, 1000, Vector2.Zero);
             ImGui.PopItemWidth();
 
-            //ImGui.PushStyleColor(ImGuiCol.Button, new Vector4(0, 120, 0, 1));
-            //ImGui.PushStyleColor(ImGuiCol.Text, new Vector4(0, 0, 204, 255));
-            if (ImGui.Button("run script") || ImGui.IsKeyPressed(0x0D))
+            if (ImGui.Button("Run") || ImGui.IsKeyPressed(0x0D))
             {
-                _scriptConsoleTextAll = string.Concat(_scriptConsoleTextAll, ">> ", _scriptConsoleText.Replace("\n","\n>> "), "\n");
-                Console.Write(string.Concat(">> ", _scriptConsoleText.Replace("\n", "\n>> "), "\n"));
+                _scriptConsoleTextAll = string.Concat(_scriptConsoleTextAll, ">> ", _scriptConsoleText.Replace("\n", "\n>> "), "\n");
                 ExecuteLuaScript(_scriptConsoleText);
             }
-            //ImGui.PopStyleColor();
-            //ImGui.PopStyleColor();
 
             ImGui.SameLine();
-            if (ImGui.Button("clear"))
+            if (ImGui.Button("Clear"))
             {
                 _scriptConsoleTextAll = "";
             }
 
             ImGui.SameLine();
             var loadPopupID = "load";
-            if (ImGui.Button("load"))
+            if (ImGui.Button("Load"))
             {
                 ImGui.OpenPopup(loadPopupID);
             }
 
             var savePopupID = "save";
             ImGui.SameLine();
-            if (ImGui.Button("save"))
+            if (ImGui.Button("Save"))
             {
                 ImGui.OpenPopup(savePopupID);
             }
@@ -66,19 +56,17 @@ namespace OpenSage.Diagnostics
             ImGui.EndChild();
             ImGui.PopStyleColor();
 
-            if (ImGui.BeginPopup("load"))
+            if (ImGui.BeginPopup(loadPopupID))
             {
-                ImGui.Text("soon");
+                ImGui.Text("Soon");
                 ImGui.EndPopup();
             }
 
-
-            if (ImGui.BeginPopup("save"))
+            if (ImGui.BeginPopup(savePopupID))
             {
-                ImGui.Text("soon");
+                ImGui.Text("Soon");
                 ImGui.EndPopup();
             }
-
         }
 
         private void ExecuteLuaScript(string userInputScript)
@@ -91,19 +79,16 @@ namespace OpenSage.Diagnostics
             }
             catch (SyntaxErrorException exeption)
             {
-                Console.WriteLine("LUA SYNTAX ERROR: ", exeption.DecoratedMessage);
                 _scriptConsoleTextAll = string.Concat(_scriptConsoleTextAll, "LUA SYNTAX ERROR: ", exeption.DecoratedMessage, "\n");
                 _consoleTextColor = new Vector4(150, 0, 0, 1);
             }
             catch (ScriptRuntimeException exeption)
             {
-                Console.WriteLine($"LUA RUNTIME ERROR: ", exeption.DecoratedMessage);
                 _scriptConsoleTextAll = string.Concat(_scriptConsoleTextAll, "LUA RUNTIME ERROR: ", exeption.DecoratedMessage, "\n");
                 _consoleTextColor = new Vector4(150, 0, 0, 1);
             }
             catch (Exception exeption)
             {
-                Console.WriteLine($"LUA RUNTIME ERROR: {0}", exeption);
                 _scriptConsoleTextAll = string.Concat(_scriptConsoleTextAll, "FATAL ERROR: ", exeption, "\n");
                 _consoleTextColor = new Vector4(150, 0, 0, 1);
             }
