+++ src/OpenSage.Game/Scripting/LuaScriptEngine.cs
@@ -1,6 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
 using MoonSharp.Interpreter;
+using OpenSage.Scripting.Lua;
 
 namespace OpenSage.Scripting
 {
@@ -12,9 +14,8 @@ namespace OpenSage.Scripting
         {
             Script.DefaultOptions.DebugPrint = text =>
             {
-                Console.WriteLine(text);
-                OpenSage.Diagnostics.LuaScriptConsole._scriptConsoleTextAll =
-                string.Concat(OpenSage.Diagnostics.LuaScriptConsole._scriptConsoleTextAll, text, "\n");
+                Diagnostics.LuaScriptConsole._scriptConsoleTextAll =
+                string.Concat(Diagnostics.LuaScriptConsole._scriptConsoleTextAll, text, "\n");
             };
 
             MainScript = new Script();
@@ -23,31 +24,31 @@ namespace OpenSage.Scripting
 
             try
             {
-                MainScript.DoString(Lua401Compatibility.Lua401CompatibilityCode);
-                var filePath = System.IO.Path.Combine(Game.ContentManager.FileSystem.RootDirectory, "Data", "Scripts", "scripts.lua");
+                LuaCompatibility.Apply(MainScript);
+
+                var filePath = Path.Combine(Game.ContentManager.FileSystem.RootDirectory, "Data", "Scripts", "scripts.lua");
                 //load scripts.lua file from folder or big file
-                if (System.IO.File.Exists(filePath))
+                if (File.Exists(filePath))
                 {
                     MainScript.DoFile(filePath);
                 }
                 else
                 {
                     //load scripts.lua from big file
-                    MainScript.DoString(Game.ContentManager.IniDataContext.GetIniFileContent(System.IO.Path.Combine("Data", "Scripts", "scripts.lua")));
+                    MainScript.DoString(Game.ContentManager.IniDataContext.GetIniFileContent(Path.Combine("Data", "Scripts", "scripts.lua")));
                 }
             }
-            catch (System.NullReferenceException) //standard case for generals and zero hour since no scripts.lua file
+            catch (NullReferenceException) //standard case for generals and zero hour since no scripts.lua file
             {
-                //Console.Write("no scripts.lua file found");
+                Debug.Write("no scripts.lua file found");
             }
-            catch (System.IO.FileNotFoundException)
+            catch (FileNotFoundException)
             {
-                Console.Write("no scripts.lua file found");
+                Debug.Write("no scripts.lua file found");
             }
             catch (Exception ex)
             {
-                //MainScript.DoString("_ALERT()");
-                Console.Write(ex);
+                Debug.Write(ex);
             }
 
             LuaEventHandlerInit();
@@ -61,17 +62,14 @@ namespace OpenSage.Scripting
             }
             catch (SyntaxErrorException exeption)
             {
-                Console.WriteLine("LUA SYNTAX ERROR: ", exeption.DecoratedMessage);
-                throw(exeption);
+                throw (exeption);
             }
             catch (ScriptRuntimeException exeption)
             {
-                Console.WriteLine($"LUA RUNTIME ERROR: ", exeption.DecoratedMessage);
                 throw (exeption);
             }
             catch (Exception exeption)
             {
-                Console.WriteLine($"LUA CRITICAL ERROR: ", exeption);
                 throw (exeption);
             }
         }
@@ -172,7 +170,7 @@ namespace OpenSage.Scripting
         public int GetInt(string number)
         {
             int i = 0;
-            if (!Int32.TryParse(number, out i))
+            if (!int.TryParse(number, out i))
             {
                 i = -1;
             }
@@ -185,11 +183,10 @@ namespace OpenSage.Scripting
             {
                 return true;
             }
-            else //if (state.Equals("0") || state.Equals("false"))
+            else
             {
                 return false;
             }
-            //return null;
         }
 
         public void AddGameObjectRefToGlobalsTable(Table gameObject)
@@ -205,13 +202,13 @@ namespace OpenSage.Scripting
 
         public string GetLuaObjectIndex(int ObjectID)
         {
-            return String.Concat("ObjID#",ObjectID.ToString("X8"));
+            return String.Concat("ObjID#", ObjectID.ToString("X8"));
         }
 
         public string Spawn(string objectType)  //quick spawn
         {
             if (objectType.Equals("")) { objectType = "AmericaVehicleDozer"; }
-            OpenSage.Logic.Object.GameObject spawnUnit = Game.Scene3D.GameObjects.Add(Game.ContentManager.IniDataContext.Objects.Find(x => x.Name == objectType), Game.Scene3D.LocalPlayer);
+            var spawnUnit = Game.Scene3D.GameObjects.Add(Game.ContentManager.IniDataContext.Objects.Find(x => x.Name == objectType), Game.Scene3D.LocalPlayer);
             var localPlayerStartPosition = Game.Scene3D.Waypoints[$"Player_{1}_Start"].Position;
             localPlayerStartPosition.Z += Game.Scene3D.Terrain.HeightMap.GetHeight(localPlayerStartPosition.X, localPlayerStartPosition.Y);
             var spawnUnitPosition = localPlayerStartPosition;
@@ -225,11 +222,11 @@ namespace OpenSage.Scripting
         public string Spawn2(string objectType, float xPos, float yPos, float zPos, float rotation)
         {
             var player = Game.Scene3D.LocalPlayer;
-            OpenSage.Logic.Object.GameObject spawnUnit = Game.Scene3D.GameObjects.Add(Game.ContentManager.IniDataContext.Objects.Find(x => x.Name == objectType), player);
+            var spawnUnit = Game.Scene3D.GameObjects.Add(Game.ContentManager.IniDataContext.Objects.Find(x => x.Name == objectType), player);
             var spawnPosition = new System.Numerics.Vector3(xPos, yPos, zPos);
             spawnPosition.Z += Game.Scene3D.Terrain.HeightMap.GetHeight(spawnPosition.X, spawnPosition.Y);
             if (zPos > spawnPosition.Z) { spawnPosition.Z = zPos; }
-            var rot = System.Numerics.Quaternion.CreateFromAxisAngle(System.Numerics.Vector3.UnitZ, OpenSage.Mathematics.MathUtility.ToRadians(rotation));
+            var rot = System.Numerics.Quaternion.CreateFromAxisAngle(System.Numerics.Vector3.UnitZ, Mathematics.MathUtility.ToRadians(rotation));
             spawnPosition += System.Numerics.Vector3.Transform(System.Numerics.Vector3.UnitX, rot);
             spawnUnit.Transform.Translation = spawnPosition;
             return GetLuaObjectIndex(Game.Scene3D.GameObjects.GetObjectId(spawnUnit));
@@ -243,7 +240,7 @@ namespace OpenSage.Scripting
             }
             else
             {
-                return System.Text.RegularExpressions.Regex.Replace(action, "([a-z])([A-Z])", "$1 $2"); 
+                return System.Text.RegularExpressions.Regex.Replace(action, "([a-z])([A-Z])", "$1 $2");
             }
         }
 
@@ -294,16 +291,16 @@ namespace OpenSage.Scripting
 
         public void ObjectSetModelCondition(string gameObject, string modelCondition)
         {
-            OpenSage.Data.Ini.Parser.IniParser Parser = new OpenSage.Data.Ini.Parser.IniParser(null, null, null, Game.SageGame);
-            Game.Scene3D.GameObjects.GetObjectById(GetLuaObjectID(gameObject)).SetModelConditionFlags(Parser.ParseEnumBitArray<OpenSage.Logic.Object.ModelConditionFlag>(modelCondition));
+            var Parser = new Data.Ini.Parser.IniParser(null, null, null, Game.SageGame);
+            Game.Scene3D.GameObjects.GetObjectById(GetLuaObjectID(gameObject)).SetModelConditionFlags(Parser.ParseEnumBitArray<Logic.Object.ModelConditionFlag>(modelCondition));
         }
 
         public bool ObjectTestModelCondition(string gameObject, string modelCondition)
         {
-            OpenSage.Data.Ini.Parser.IniParser Parser = new OpenSage.Data.Ini.Parser.IniParser(null, null, null, Game.SageGame);
-            var modelConditionBitArray = Parser.ParseEnumBitArray<OpenSage.Logic.Object.ModelConditionFlag>(modelCondition);
+            var Parser = new Data.Ini.Parser.IniParser(null, null, null, Game.SageGame);
+            var modelConditionBitArray = Parser.ParseEnumBitArray<Logic.Object.ModelConditionFlag>(modelCondition);
             var modelconditionBitArrayEnum = Game.Scene3D.GameObjects.GetObjectById(GetLuaObjectID(gameObject)).ModelConditionStates;
-            foreach (OpenSage.Data.Ini.BitArray<OpenSage.Logic.Object.ModelConditionFlag> i in modelconditionBitArrayEnum)
+            foreach (var i in modelconditionBitArrayEnum)
             {
                 if (i == modelConditionBitArray)
                 {
@@ -315,10 +312,10 @@ namespace OpenSage.Scripting
 
         public void ObjectClearModelCondition(string gameObject, string modelCondition)
         {
-            OpenSage.Data.Ini.Parser.IniParser Parser = new OpenSage.Data.Ini.Parser.IniParser(null, null, null, Game.SageGame);
-            var modelConditionBitArray = Parser.ParseEnumBitArray<OpenSage.Logic.Object.ModelConditionFlag>(modelCondition);
+            var Parser = new Data.Ini.Parser.IniParser(null, null, null, Game.SageGame);
+            var modelConditionBitArray = Parser.ParseEnumBitArray<Logic.Object.ModelConditionFlag>(modelCondition);
             var modelconditionBitArrayEnum = Game.Scene3D.GameObjects.GetObjectById(GetLuaObjectID(gameObject)).ModelConditionStates;
-            foreach (OpenSage.Data.Ini.BitArray<OpenSage.Logic.Object.ModelConditionFlag> i in modelconditionBitArrayEnum)
+            foreach (var i in modelconditionBitArrayEnum)
             {
                 if (i == modelConditionBitArray)
                 {
@@ -345,7 +342,7 @@ namespace OpenSage.Scripting
         public void ObjectForbidPlayerCommands(string gameObject, string state)
         {
         }
-        
+
         public void ObjectSetGeometryActive(string gameObject, string geometryName, string state)
         {
         }
@@ -366,7 +363,7 @@ namespace OpenSage.Scripting
 
         public int ObjectGetHealthFraction(string gameObject)
         {
-            return (int)Game.Scene3D.GameObjects.GetObjectById(GetLuaObjectID(gameObject)).Health;
+            return (int) Game.Scene3D.GameObjects.GetObjectById(GetLuaObjectID(gameObject)).Health;
         }
 
         public string ObjectDescription(string gameObject)  //EXAMPLE C&C3: "Object 1187 (_jIWv4) [NODAvatar, owned by player 3 (MetaIdea)]"
@@ -408,7 +405,7 @@ namespace OpenSage.Scripting
         }
 
         public void HordeBroadcastEventToMembers(string gameObject, string eventName) //unknown params
-        {   
+        {
         }
 
         public void ObjectBroadcastEventToEnemies(string gameObject, string eventName, string radius)
@@ -432,7 +429,7 @@ namespace OpenSage.Scripting
         }
 
         public void ObjectSetFearFactor(string gameObject, string FearFactor) //unknown params
-        {   
+        {
         }
 
         public bool ObjectTestCanSufferFear(string gameObject)
@@ -466,247 +463,13 @@ namespace OpenSage.Scripting
 
         public double GetRandomNumber()  //attention for multiplayer sync
         {
-            Random random = new Random();
+            var random = new Random();
             return random.NextDouble();
         }
 
         public int GetFrame()
         {
-            return (int)Game.CurrentFrame;
+            return (int) Game.CurrentFrame;
         }
-    }
-
-    public class LuaScriptEngineObject : GameSystem //lua environment instance for each object with lua registration
-    {
-        public Script MainScript { get; set; }
-
-        public LuaScriptEngineObject(Game game) : base(game)
-        {
-            Script.DefaultOptions.DebugPrint = text => Console.WriteLine(text);
-            MainScript = new Script();
-            FunctionInit();
-            MainScript.DoString(Lua401Compatibility.Lua401CompatibilityCode);
-        }
-
-        public void FunctionInit()
-        {
-            MainScript.Globals["CurDrawableObjectStatus"] = (Func<string, bool>) CurDrawableObjectStatus;
-            MainScript.Globals["CurDrawableModelcondition"] = (Func<string, bool>) CurDrawableModelcondition;
-            MainScript.Globals["CurDrawableGetCurrentTargetBearing"] = (Func<float>) CurDrawableGetCurrentTargetBearing;
-            MainScript.Globals["CurDrawableGetCurrentTargetHeight"] = (Func<float>) CurDrawableGetCurrentTargetHeight;
-            MainScript.Globals["CurDrawableGetCurrentTargetDistance"] = (Func<float>) CurDrawableGetCurrentTargetDistance;
-            MainScript.Globals["CurDrawableIsCurrentTargetKindof"] = (Func<string, bool>) CurDrawableIsCurrentTargetKindof;
-            MainScript.Globals["CurDrawablePrevAnimation"] = (Func<string>) CurDrawablePrevAnimation;
-            MainScript.Globals["CurDrawablePrevAnimationstate"] = (Func<string>) CurDrawablePrevAnimationstate;
-            MainScript.Globals["CurDrawablePrevAnimFraction"] = (Func<float>) CurDrawablePrevAnimFraction;
-            MainScript.Globals["GetClientRandomNumberReal"] = (Func<int, int, float>) GetClientRandomNumberReal;
-            MainScript.Globals["GetFrame"] = (Func<int>) GetFrame;
-            MainScript.Globals["CurDrawablePlaySound"] = (Action<string>) CurDrawablePlaySound;
-            Action CurDrawableAllowToContinueAction = () => CurDrawableAllowToContinue();
-            MainScript.Globals["CurDrawableAllowToContinue"] = CurDrawableAllowToContinueAction;
-            MainScript.Globals["CurDrawableSetTransitionAnimstate"] = (Action<string>) CurDrawableSetTransitionAnimstate;
-            MainScript.Globals["CurDrawableShowModule"] = (Action<string>) CurDrawableShowModule;
-            MainScript.Globals["CurDrawableHideModule"] = (Action<string>) CurDrawableHideModule;
-            MainScript.Globals["CurDrawableHideSubObjectPermanently"] = (Action<string>) CurDrawableHideSubObjectPermanently;
-            MainScript.Globals["CurDrawableShowSubObjectPermanently"] = (Action<string>) CurDrawableShowSubObjectPermanently;
-            MainScript.Globals["CurDrawableShowSubObject"] = (Action<string>) CurDrawableShowSubObject;
-            MainScript.Globals["CurDrawableHideSubObject"] = (Action<string>) CurDrawableHideSubObject;
-        }
-
-        public void LuaEventHandlerInit()
-        {
-        }
-
-        public void LuaEventHandler()
-        {
-        }
-
-            public bool CurDrawableObjectStatus(string objectStatus)
-        {
-            return true;
-        }
-
-        public bool CurDrawableModelcondition(string modelCondition)
-        {
-            return true;
-        }
-
-        public float CurDrawableGetCurrentTargetBearing() //example return 0.10
-        {
-            return 0;
-        }
-
-        public float CurDrawableGetCurrentTargetHeight()
-        {
-            return 0;
-        }
-
-        public float CurDrawableGetCurrentTargetDistance()
-        {
-            return 0;
-        }
-
-        public bool CurDrawableIsCurrentTargetKindof(string kindof)
-        {
-            return true;
-        }
-
-        public string CurDrawablePrevAnimation() //example return "runtrans_bs"
-        {
-            return "";
-        }
-
-        public string CurDrawablePrevAnimationstate() //example return "Moving_Sword"
-        {
-            return "";
-        }
-
-        public float CurDrawablePrevAnimFraction() //example return 0.66
-        {
-            return 0;
-        }
-
-        public float GetClientRandomNumberReal(int low, int high)
-        {
-            return 0;
-        }
-
-        public int GetFrame()
-        {
-            return 0;
-        }
-
-        public void CurDrawablePlaySound(string soundName)
-        {
-
-        }
-
-        public void CurDrawableAllowToContinue()
-        {
-
-        }
-
-        public void CurDrawableSetTransitionAnimstate(string transitionstateName)
-        {
-
-        }
-
-        public void CurDrawableShowModule(string moduleTagName)  //example input "ModuleTag_DrawLight"
-        {
-
-        }
-
-        public void CurDrawableHideModule(string moduleTagName)
-        {
-
-        }
-
-        public void CurDrawableHideSubObjectPermanently(string subObject)
-        {
-
-        }
-
-        public void CurDrawableShowSubObjectPermanently(string subObject)
-        {
-
-        }
-
-        public void CurDrawableShowSubObject(string subObject)
-        {
-
-        }
-
-        public void CurDrawableHideSubObject(string subObject)
-        {
-
-        }
-    }
-
-    public class Lua401Compatibility
-    {
-        public const string Lua401CompatibilityCode = @"
-                globals = _G
-                function getn(table) return #table end
-                closefile = io.close
-                flush = io.flush
-                openfile = io.open
-                read = io.read
-                tmpname = os.tmpname
-                write = io.write
-                abs = math.abs
-                acos = math.acos
-                asin = math.asin
-                atan = math.atan
-                atan2 = math.atan2
-                ceil = math.ceil
-                cos = math.cos
-                cosh = math.cosh
-                deg = math.deg
-                exp = math.exp
-                floor = math.floor
-                mod = math.fmod
-                mod2 = math.modf
-                frexp = math.frexp
-                ldexp = math.ldexp
-                log = math.log
-                max = math.max
-                min = math.min
-                PI = math.pi
-                randomseed = math.randomseed
-                rad = math.rad
-                random = math.random
-                sin = math.sin
-                sqrt = math.sqrt
-                tan = math.tan
-                clock = os.clock
-                date = os.date
-                execute = os.execute
-                exit = os.exit
-                getenv = os.getenv
-                remove = os.remove
-                rename = os.rename
-                setlocale = os.setlocale
-                strbyte = string.byte
-                strchar = string.char
-                strfind = string.find
-                format = string.format
-                gsub = string.gsub
-                strlen = string.len
-                strlower = string.lower
-                strrep = string.rep
-                strsub = string.sub
-                strupper = string.upper
-                tinsert = table.insert
-                tremove = table.remove
-                sort = table.sort
-                function log10(number) return math.log(number,10) end
-                call = pcall
-                function seek(filehandle, whence, offset) return filehandle:seek(whence, offset) end
-                rawgettable = rawget
-                rawsettable = rawset
-                function getglobal(index) return _G[index] end
-                function setglobal(index, value) _G[index]=value end
-                function foreach(t, f)
-                    for i, v in t do
-                        local res = f(i, v)
-                        if res then return res end
-                    end
-                end
-                function foreachi(t, f)
-                    for i=1,#(t) do
-                        local res = f(i, t[i])
-                        if res then return res end
-                    end
-                end
-                function readfrom(file) _INPUT = io.open(file,'r') return io.read(file) end
-                function writeto(file) _OUTPUT = io.open(file,'w+') return io.output(file) end
-                function appendto(file) _OUTPUT = io.open(file,'a+') return io.output(file) end
-                --debug = debug.debug
-                function rawgetglobal(index) rawget(_G, index) end
-                function rawsetglobal(index, value) return rawset(_G, index, value) end
-                function _ALERT(...) print('error') end
-                function _ERRORMESSAGE(...) print('error') end";
-                //unsupported (and never used in any SAGE game and it's mods): debug and tag methods, %upvalues
-                //DEPRECATED and actually removed from original SAGE: rawgetglobal, rawsetglobal, foreachvar, nextvar
     }
 }
